Refactor and overhaul internals of library
While the public interfaces of the library remain compatible, the implementation is
now heavily refactored to decouple the toolkit from a concrete PDF library.

Additional wrappers for PDF libraries can now implement the WrapperInterface, and in
addition BinaryPathAwareInterface if they allow to manually set their binary path.
See PdftkWrapper as an example implementation for pdftk.

The testability is now greatly improved by this decoupling. Furthermore, the test
suite is now featuring many unit- and functional tests, with the latter annotated
with @group FunctionalTest.

For common tasks with PDFs and easier tests, constants for page sizes and orientations
are now provided.

Possible BC breaks:
- The methods Bookmarks::importFromDump(), Pages::importFromDump() and Metadata::importFromDump()
are removed. They were considered internal implementation details and were coupled to pdftk.